Move ResponseNormalizer
